Add files via upload
The electrical method is one of the important methods in Geo Physical Imaging in which measures the resistivity of the sub-surface of the ground by injecting a current into the ground and measuring the voltage drop between two electrodes. Since the ground is not a homogeneous medium, the resistivity which we measure is not a true resistivity but it is an apparent resistivity.  I, firstly, defined a model by #Res2Dmod software and then estimated the sub-surface. Imported the result into #Res2Dinv and estimated the inversion model. Repeat the iteration until reaching an acceptable MMSE between the imported model and the real model.